Move IModeratorController to common package
To maintain network compatibility, it appears only the remote name
cannot change, but the fully-qualified name of the remote type is
free to change (as long as it maintains the same methods and
signatures).
