+++ game-core/src/main/java/games/strategy/engine/lobby/client/LobbyClient.java
@@ -1,7 +1,7 @@
 package games.strategy.engine.lobby.client;
 
+import games.strategy.engine.lobby.common.IModeratorController;
 import games.strategy.engine.lobby.common.LobbyConstants;
-import games.strategy.engine.lobby.server.IModeratorController;
 import games.strategy.engine.message.IChannelMessenger;
 import games.strategy.engine.message.IRemoteMessenger;
 import games.strategy.net.IMessenger;
