+++ game-core/src/main/java/games/strategy/engine/lobby/client/ui/LobbyGamePanel.java
@@ -25,9 +25,9 @@ import javax.swing.SwingUtilities;
 import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.startup.ui.InGameLobbyWatcher;
 import games.strategy.engine.framework.startup.ui.ServerOptions;
+import games.strategy.engine.lobby.common.IModeratorController;
 import games.strategy.engine.lobby.common.LobbyConstants;
 import games.strategy.engine.lobby.server.GameDescription;
-import games.strategy.engine.lobby.server.IModeratorController;
 import games.strategy.net.INode;
 import games.strategy.net.Messengers;
 import games.strategy.net.Node;
