+++ game-core/src/main/java/games/strategy/triplea/ui/menubar/LobbyMenu.java
@@ -28,8 +28,8 @@ import games.strategy.engine.lobby.client.login.LobbyLoginPreferences;
 import games.strategy.engine.lobby.client.ui.LobbyFrame;
 import games.strategy.engine.lobby.client.ui.MacLobbyWrapper;
 import games.strategy.engine.lobby.client.ui.TimespanDialog;
+import games.strategy.engine.lobby.common.IModeratorController;
 import games.strategy.engine.lobby.common.LobbyConstants;
-import games.strategy.engine.lobby.server.IModeratorController;
 import games.strategy.engine.lobby.server.IUserManager;
 import games.strategy.engine.lobby.server.login.RsaAuthenticator;
 import games.strategy.engine.lobby.server.userDB.DBUser;
